move cmd related utils to a separate file
